Dedent() removes surrounding newlines
The most common use-case for Dedent() is a situation like:

    multiLineString = dedent.Dedent(`
        This is a
        multi-line string.
    `)

Most programmers would reasonably expect multiLineString to be equal to
`"This is a\nmulti-line string." Currently Dedent() returns `"\nThis is
a\nmulti-line string.\n". When using raw string syntax, there is currently no good way to specify a multi-line string without a "\n". This commit strips a "\n" prefix and suffix, only if the string has both a "\n" prefix and a "\n" suffix.